More re-jiggering of s3 tables and index
