+++ datacube/index/metadata_types.py
@@ -1,3 +1,5 @@
+from __future__ import absolute_import
+
 import logging
 import warnings
