+++ datacube/index/postgres/_connections.py
@@ -168,13 +168,13 @@ class PostgresDb(object):
             _LOG.warning('Application name is too long: Truncating to %s chars', (64 - len(_LIB_ID) - 1))
         return full_name[-64:]
 
-    def init(self, with_permissions=True, with_s3_tables=False):
+    def init(self, with_permissions=True):
         """
         Init a new database (if not already set up).
 
         :return: If it was newly created.
         """
-        is_new = tables.ensure_db(self._engine, with_permissions=with_permissions, with_s3_tables=with_s3_tables)
+        is_new = tables.ensure_db(self._engine, with_permissions=with_permissions)
         if not is_new:
             tables.update_schema(self._engine)
