+++ datacube/index/postgres/tables/_core.py
@@ -25,7 +25,6 @@ SQL_NAMING_CONVENTIONS = {
 }
 SCHEMA_NAME = 'agdc'
 METADATA = MetaData(naming_convention=SQL_NAMING_CONVENTIONS, schema=SCHEMA_NAME)
-S3_METADATA = MetaData(naming_convention=SQL_NAMING_CONVENTIONS, schema=SCHEMA_NAME)
 
 _LOG = logging.getLogger(__name__)
 
@@ -43,7 +42,7 @@ def _get_quoted_connection_info(connection):
     return db, user
 
 
-def ensure_db(engine, with_permissions=True, with_s3_tables=False):
+def ensure_db(engine, with_permissions=True):
     """
     Initialise the db if needed.
     """
@@ -75,8 +74,6 @@ def ensure_db(engine, with_permissions=True, with_s3_tables=False):
             _LOG.info('Creating tables.')
             c.execute(TYPES_INIT_SQL)
             METADATA.create_all(c)
-            if with_s3_tables:
-                S3_METADATA.create_all(c)
             c.execute('commit')
         except:
             c.execute('rollback')
