+++ datacube/index/products.py
@@ -1,3 +1,5 @@
+from __future__ import absolute_import
+
 import logging
 
 from cachetools.func import lru_cache
